Remove redundant "Show" link from tables
We have a snippet of javascript that makes table rows clickable,
and links each table rows to the resource show page.

Because of this, the "show" link is redundant.
The "show" link was originally added over concerns about accessibility.
However, according to a survey of screen readers and browsers,
javascript is nearly ubiquitous.
We don't need to worry about browsers that don't have it enabled.

More information at:
http://webaim.org/projects/screenreadersurvey5/#javascript

- Add type class to table cells based on attr type

This helps us identify areas that are safe to click in specs,
and will let us style fields differently depending on their type.

- Add news for the 0.0.4 release