Bump Sass version ~> 3.4
After installing Administrate on Upcase (which was using Sass v3.2.19),
I got errors concerning CSS selector syntax that was introduced
in Sass 3.3.0, and about using the `&` parent selector in SassScript,
which was introduced in `3.4.0`.

Instead of relying on the indirect `sass` dependency through `neat`,
we'll be explicit about requiring `sass ~> 3.4`.

We'll use appraisal to test against different gem versions

http://sass-lang.com/documentation/file.SASS_CHANGELOG.html#340_18_august_2014