Change Dashboard API to use constants over methods
Why?

It is unlikely that a developer would want to return non-constant data
for any of the API methods in DashboardManifest or `ModelDashboard`
classes.

See relevant discussion on
[the pull request to add Administrate to Hound][1].

[1]: https://github.com/thoughtbot/hound/pull/897

- - -

Related improvement:
Make generator specs actually read generated file

Why?

It was possible that the generated files contained the strings we were
looking for, but not in the correct order or context.

By actually reading the generated files, we can be sure that the
dashboard is semantically what we want it to be.

This commit only changes one test - other tests should be updated as
they're touched.