Support custom class names for `belongs_to`
Problem:

When a `belongs_to` relationship was defined with a custom `class_name`,
Administrate wouldn't pick up on the option and would try to fetch the
object using the relationship name. This resulted in 500 errors, most
commonly on form pages.

See https://trello.com/c/hkZynxwL

Solution:

Let the user pass a `class_name` option to the Administrate
field to override the default class lookup.

The solution is essentially what we did for `has_many` relationships
back in c4aa44d916.