Add a generator for new field types
According to documentation at:
https://administrate-docs.herokuapp.com/adding_custom_field_types

Problem:

One of the key features of Administrate is the ability to add custom
fields for different attribute types.
At the moment, developers need to add fields that fit a certain API,
which requires them to dig into the Administrate source code
to fit the correct API.

Solution:

Add a generator for easily generating new field types.