Update datetime_picker_rails to fix asset bug
https://trello.com/c/C52Z0ipx/87-fix-asset-initializers

Problem:

Under some circumstances (reported by @mcmire),
host applications would not inherit
the asset configuration in `config/initializers/assets.rb`,
which adds several font assets to the precompilation path.

Without those assets in the path,
Rails is not able to properly serve up the assets,
and this presents to the user as a 500 error on any admin page.

Solution:

Version 0.0.5 of the `datetime_picker_rails` gem
moves the font assets to the `app` directory,
which is part of Rails' default asset precompilation path.

This makes the extra configuration lines unnecessary,
thereby removing the source of the bug.