Support translations for ActiveRecord models
Problem:

Developers are unable to display their ActiveRecord classes in the UI
by any other name than the titleized class name.

For example,
if a team is building an admin dashboard that will be used in Russian,
they may have ActiveRecord models with English names, that need to be
translated before they're displayed in the UI.

Solution:

Rails has support for [translating ActiveRecord Models][1],
which relies on defining I18n translations
under the `activerecord.models` namespace.

Rails expects translations in the form:

```yaml
en:
  activerecord:
    models:
      user:
        one: Customer
        other: Customers
```

The separate `one` and `other` keys allow Rails to display
the correct plural or singular form of the word,
depending on the context.

We've used Rails' built-in `Foo.model_name.human` method
to correctly pull the translations out of I18n,
with sane fallbacks if the translation is not defined.

[1]: http://guides.rubyonrails.org/i18n.html#translations-for-active-record-models