Set up visual regression testing with Percy
See https://percy.io/

Problem:

For pull requests with UI changes, it can be difficult to tell if the
changes affect elements on other pages.
It can also be tedious to pull the changes down to a local machine
if the contributor has forgotten to add screenshots.

Solution:

Add visual regression snapshots to our test suite with Percy.
After we record the inital set of screenshots,
future test suite runs will be compared against them
to recognize and call out any changes.