Remove placeholder styles
This was only used for slightly changing the text color of the global search
placeholder text. But the CSS output was much larger and applied to all text
inputs. The gain was so slight that removing it for now is a good path forward.

This also removes one more Bourbon mixin use.