Add release notes & release v `0.2.0`
Problem:

This repository contains both the administrate gem
and the demo application.
The demo application uses the image field,
which means that its Gemfile will be dependent on the new
administrate-field-image gem.

In turn, the administrate-field-image gem
is dependent on the administrate gem in this repository.
This creates a little bit of a chicken-and-egg problem
for publishing both of those gems.

Solution:

Our release plan is as follows:

- Release this branch as a new version of the administrate gem,
  as version `0.2.0.rc1`
- Make administrate-field-image dependent on
  `"administrate", "0.2.0.rc1", "~> 0.2.0"`
- Release administrate-field-image v `0.0.1`
- Make the demo app's Gemfile dependent on
  `"administrate-field-image", "~> 0.0"`
- Re-push this branch to Github to run CircleCI suite. All of the gems
  would be available on rubygems.org, so the suite should pass.
- Update and release this branch as version `0.2.0`
- Merge branch