Fix inferred path issue w/name-spaced resources
If a name-spaced model and corresponding dashboard exists we will run
into a problem. When we do something like link to a resource's `new`
page, we currently use `page.resource_name` to infer what the path to
that page should be. This probably works 90+ percent of the time because
most models are not name-spaced. That is until you get into apps that
either mount engines containing their own models (and are appropriately
name-spaced), or larger apps that namespace conceptually related
resources.

For purposes of this example, and the most recent issue that I've run
into, I have a `Logged` namespace for both a model and a dashboard.
Let's say we have a "Logged::Order" model/dashboard. What happens?

```
link_to("New Logged Order", [:new, namespace, page.resource_name])
=> 'admin_logged/order_path'
```

That path won't fly. Instead, we want to see `'admin_logged_order_path'`.

As a result, I've added a `resource_path` method to the base page class
that will handle this responsibility, and backed the original and new
method with some tests to make sure they are behaving as expected.

... more to come related to similar issue in
`./app/views/administrate/application/_sidebar.html.erb`