Change sidebar partial to work w/namespaced models
The sidebar uses the resource for both the *name* of the link, and the
inferred path that we're linking to. The problem with this comes when
the resource/model/dashboard is namespaced. It will break when it runs
into something that ends up looking like:

```
link_to "logged customer", [:admin, "logged/customer"]
```

... which will break. Instead, it should turn out to look something like:

```
link_to "logged customer", [:admin, "logged_customer"]
```

This commit fixes this by updating the resources being looped through
from strings (for the name), to an object exposing both a name and a
path, (and corresponding strings and symbols).