Render all records allowed by the authorization scope (#1596)
On a collection page with four items, the cells show, but the data doesn't.

The Pundit scope for the model was returning the records, but the per-resource
policy was returning false for show? on each one of them.

This is a bit of an edge case: if the user can't see the resources, surely the
scope shouldn't be returning them. Still I can think of a use case where
some users can see them listed (with limited information) but can't access
the show page (with additional information).