Use Rails defaults from v6.0.3.2 for gitignore (#1726)
When #1718 was merged, it introduced a potential solution to /tmp/pids/.keep
being unintentionally deleted. Unfortunately this didn't fix the issue and in fact
seems to have made it worse.

The issue may have been solved already, and it was only surfacing again
when moving between commits that included #cea66f8 and those that didn't.

The changes I'm proposing to gitignore are as currently generated in new Rails
apps.