Fix pagination of "Page" models (#1725)
Using the param :page for top-level pagination (ie: index pages as opposed to
has_many lists in show pages) conflicts with paginating resources whose type
happens to be "Page".

Instead of a unit test, a feature spec is used as the pagination code is all over
the place making this quite tricky. Ideally, we should move all the pagination
code into a single module that can be tested in isolation. This in turn is not
easy, as Kaminari doesn't make it easy.

A new feature spec called pagination_spec.rb, bringing over other pre-existing
pagination examples, to help organise things, and also to take advantage of
the helper method expect_to_appear_in_order in the new example.

Fixes #1709