Correct "required" asterisk when using validation option `:on` (#1872)
We show a "required" asterisk when fields are required. However there are
some edge cases, like when the requirement has an if/unless. We don't
consider those actually required and we don't show the asterisk there.

But there are other edge cases. Issue #1860 shows how the
validation option :on can also create cases where an apparent 
requirement is not actually a requirement. Normally this option has 
the values :create or :update, but there are other possibilities.

This PR tries to solve the issue. There are three commits, each an 
incremental improvement:

1. Consider :on to be the same as :if and :unless, and hide the asterisk.
2. Refactor specs to avoid the mystery guests involved in the
  examples for :if, :unless, and :on. This also allows for more scalable
  specs, should we decide to add more fine-grained handling of these
  options.
3. Consider the special (and most common) cases of on: :create
  and on: :update, and hide the asterisk for any other :on value.

Fixes #1860