add params to collection filter (#2005)
* add params to collection filter

To filter a collection you can define COLLECTION_FILTER and use them like "filter:".

With this PR you can also params to those filter and use them like "filter:param". The COLLECTION_FILTER will be called with the params a second argument.

This ideas was already mentioned in the initial PR for COLLECTION_FILTERS, but not implemented. (github.com/thoughtbot/administrate/pull/947)

* make sure filter and terms get not confused

a filter like kind:vip and a terms like order:12 should not get confused.

* take care of hound warnings

* 2 more hound warnings taken care of