Enable ordering the BelongsTo fields by using `order` option.
The change let the gem users to be able to order the fields by
the column they'll specify, instead of ordering by `id`.
I've also updated the query to solve the possible issues with
`PG::AmbiguousColumn` errors - for now it's not handled in
any way, so if the user is using the tables that use the same
columns names, then the error will apear and the only way to fix
that is to monkey patch the gem. To solve that, I've added
the column name as the prefix for both possible forms of the query.