Migrate JavaScript to integrate with `importmap-rails`
Depend on [importmap-rails][] to serve our client-side JavaScript.

As part of this change, move the `render "javascript"` partial from the
bottom of the `<body>` element to the bottom of the `<head>` element.

Replace a server-side dependency on `jquery-rails` with a client-side
importmaps declaration for `jquery` and `jquery-ujs`. We're unable to do
the same with `selectize` because we're depending on that pre-bundled
package to expose CSS.

[importmap-rails]: https://github.com/rails/importmap-rails/