Re-bundle JS in a fresh environment (#2710)
`npm`/`yarn` is non-deterministic when it comes to bundling
dependencies, and so you can very easily end up with a dependency drift
(especially when you're being strict about what you build elsewhere).

It turns out that this is what #2680 was showing. If we clean the whole
environment (`git clean -fdx`), then run `setup` and the other steps
that the Action runs through, we get the same result.