Run diff-check on JS/CSS to catch updates (#2680)
In #2673, we discussed catching changes in the `builds` directory as
being quite annoying in development. This can happen when people
contribute changes (and should probably also commit their build changes
to make future development easier), plus also when dependencies change.

Catching this with `diff-check` should mean that we see this at the
point of introduction, rather than later on.