Introduce --main-content-shadow to work around sass
sass has a function called [rgb], and so does [css]. If I try to use
the css rgb in a sass context, sass tries to use it first. If I'm using
the [relative color syntax], sass complains that I'm passing too many
arguments. One way around this is to use a css property to define
whatever I need and then use that instead.

[rgb]: https://sass-lang.com/documentation/modules/#rgb
[css]: https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb
[relative color syntax]: https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb#using_relative_colors_with_rgb