Switch to using the feather icon set (#2809)
In implementing #2758, we realised that there's no citation of the
source of the icons. It's quite likely that they were custom made at the
time, but if they weren't then this was a mistaken license violation.
Most practically though, it's a problem when we need to add further
icons, as we'd want to use a set which has plenty of options for new
features.

Feather is a nice set of icons, covering everything we could ever likely
need, is MIT licensed and fits well with our current style.

This replaces the existing icons in place with the new set, using
`currentColor` to provide a path to override the colour in a standard
manner (this means that you set the CSS `color` properly and don't need
to change `fill` or `stroke`.)

https://feathericons.com/
https://github.com/feathericons/feather