Restrict attributes shown on the Customer index (#2810)
As we've added more attributes to the `Customer`, the index has gotten
wider and wider until the page layout has broken. It's the first page
you land on when looking at the demo, so in addition to setting a bad
first impression, it's not reflective of what someone would setup
themselves.

We fix this by reducing the attributes shown to be sensible, but avoid
breaking the page layout.