Replace form_for with form_with (#2825)
`form_for` was deprecated in Rails 5.1, but `form_with` is a close
drop-in replacement.

In the demo app, we need to set
`config.action_view.form_with_generates_ids = true` to automatically
generate `id` and `for` attributes.  However, this is already `true` by
default in `load_defaults` from Rails 6.0 onward, so it shouldn’t cause
problems in most applications.