Fix broken Selectize on polymorphic select field (#2839)
When building the select field in the polymorphic form, we omitted to
pass `nil` to the choices argument, which resulted in an erroneous
argument order, and a {} was passed to it instead.
This caused the selectize gem to fail applying the `.selectize` method
under the hood. We introduced this in #2447.

Passing `nil` will now restore its intended behavior and allow selectize
to perform the dynamic seach in the select field again.

To ensure we don't hit this problem again, we implement some feature
tests which specifically trigger the Selectize functionality.

Fixes #2821