Added a label to the index table's "actions" columns
The "actions" columns in the collection table did not have any text in the column header cells. This was fine visually because it's a common pattern to use the final cells for action buttons. But people using a screen readers may not have this visual context. By adding a title of "Actions" to the column, this title is now announced by screen readers when moving into that column.